+++ src/main/java/de/blablubbabc/insigns/InSigns.java
@@ -4,9 +4,11 @@
  */
 package de.blablubbabc.insigns;
 
+import java.util.List;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
+import org.bukkit.block.Sign;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -25,6 +27,9 @@ public class InSigns extends JavaPlugin {
 
 		// Setup metrics:
 		this.setupMetrics();
+
+		// Update the nearby signs of all already online players:
+		this.updateNearbySignsOfOnlinePlayersDelayed();
 	}
 
 	private void setupMetrics() {
@@ -33,12 +38,39 @@ public class InSigns extends JavaPlugin {
 		}
 	}
 
+	private void updateNearbySignsOfOnlinePlayersDelayed() {
+		int delayTicks = this.getPlayerJoinSignUpdateDelay();
+		if (delayTicks <= 0) return; // Disabled
+
+		Bukkit.getScheduler().runTaskLater(this, () -> {
+			for (Player player : Bukkit.getOnlinePlayers()) {
+				this.updateNearbySigns(player);
+			}
+		}, delayTicks);
+	}
+
 	@Override
 	public void onDisable() {
 	}
 
 	public int getPlayerJoinSignUpdateDelay() {
 		return this.getConfig().getInt("player-join-sign-update-delay", 2);
+	}
+
+	void updateNearbySigns(Player player) {
+		assert player != null && player.isOnline();
+		List<Sign> nearbySigns = Utils.getNearbyTileEntities(player.getLocation(), Bukkit.getViewDistance(), Sign.class);
+		for (Sign sign : nearbySigns) {
+			Utils.sendSignUpdate(player, sign);
+		}
+	}
+
+	void updateNearbySignsDelayed(Player player, int delayTicks) {
+		if (delayTicks <= 0) return; // Disabled
+		Bukkit.getScheduler().runTaskLater(this, () -> {
+			if (!player.isOnline()) return;
+			updateNearbySigns(player);
+		}, delayTicks);
 	}
 
 	SignSendEvent callSignSendEvent(Player player, Location location, String[] rawLines) {